requirements.txt
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
schedule==1.2.1
lxml==4.9.3
